feat: add appointment CRUD

+++ src/StartupService.cs
@@ -1,4 +1,7 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using AppointmentScheduler.Models;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Storage;
 
 namespace AppointmentScheduler
 {
@@ -14,8 +17,23 @@ namespace AppointmentScheduler
         public async Task StartAsync(CancellationToken cancellationToken)
         {
             using var dbContext = _dbContextFactory.CreateDbContext();
+            var wasExist = dbContext.Database.GetService<IRelationalDatabaseCreator>().Exists();
+
+            if (wasExist) return;
+
             await dbContext.Database.EnsureDeletedAsync();
             await dbContext.Database.EnsureCreatedAsync();
+
+            var userAccount = new UserAccount
+            {
+                Username = "vintorezsr",
+                Password = BCrypt.Net.BCrypt.HashPassword("123456789"),
+                Email = "vintorezsr@gmail.com"
+            };
+
+            dbContext.UserAccounts?.Add(userAccount);
+
+            await dbContext.SaveChangesAsync();
         }
 
         public Task StopAsync(CancellationToken cancellationToken)
